chore: move network display to correct location
